daburu hort feesu

+++ codeine_django/organization/views_contribution_payment.py
@@ -30,9 +30,9 @@ def contribution_payment_view(request):
 
         with transaction.atomic():
             try:
-                month_duration = float(data['month_duration'])
+                month_duration = int(data['month_duration'])
                 payment_transaction = PaymentTransaction(
-                    payment_amount=float(data['contribution']) * month_duration,
+                    payment_amount=(float(data['contribution']) * month_duration),
                     payment_type=data['payment_type']
                 )
                 payment_transaction.save()
